requirements.txt
streamlit==1.28.0
Pillow==10.0.0
imageio==2.31.1
imageio-ffmpeg==0.4.8
numpy==1.24.0 